requirements.txt
anytree==2.4.3
Django==2.0.6
python-irodsclient==0.8.0

mysqlclient

celery==4.2
django-celery-results==1.0.1
django-celery-beat==1.1.0

django-suit==0.2.26


djangosaml2==0.16.11

Pillow==5.0.0

Wand==0.4.4

reportlab==3.4.0






